Render Multiple Objects, rotate and launch arrows
change the program so that it iterates through an object array when rendering, drawing each object. These objects are added to the array when loaded. Arrows can be rotated to extend using the directional keys, and launched at a set velocity using the space bar. This causes a new arrow to be loaded.